app/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
PyJWT==1.7.1
pytz==2022.1
psycopg2
django-cors-headers
sqlparse==0.4.2
